cut and paste the schedule of  location inside the location scouting
